archive/untitled folder/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
geopandas>=0.14.0
h3>=3.7.6
shapely>=2.0.0
pyproj>=3.6.0

# Data processing
pyarrow>=14.0.0  # For parquet caching
requests>=2.31.0
pyyaml>=6.0

# Climate modeling
earth2studio>=0.1.0
torch>=2.0.0
xarray>=2023.0.0
netcdf4>=1.6.0

# Remote sensing / Earth Engine (optional)
earthengine-api>=0.1.350
rasterio>=1.3.0
rioxarray>=0.15.0

# Visualization
folium>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Type checking (optional)
mypy>=1.5.0
types-requests>=2.31.0
types-pyyaml>=6.0.0
